services/recommendation-management/requirements.txt
openai==1.70.0
pika==1.3.1
requests==2.28.2
python-dotenv==1.0.0